refactor: break circular dependencies by replacing barrel back-imports with direct source imports
Four internal modules in packages/shared imported back from the barrel
(../index.js), making index.ts participate in ~36 import cycles. Redirect
each back-import to its direct source module so index.ts no longer takes
part in any cycle. Public API of index.ts is unchanged.
