+++ packages/shared/lib/services/environment.service.ts
@@ -4,16 +4,18 @@ import db from '@nangohq/database';
 import { report, useLambdaKeepWarm } from '@nangohq/utils';
 
 import { PROD_ENVIRONMENT_NAME } from '../constants.js';
-import { configService, externalWebhookService, getGlobalOAuthCallbackUrl } from '../index.js';
+import configService from './config.service.js';
 import customerKeyService from './customerKey.service.js';
+import * as externalWebhookService from './external-webhook.service.js';
 import { pubsub } from '../utils/pubsub.js';
 import { getPlan, lambdaKeepWarmProvisionedConcurrencyMultiplier } from './plans/plans.js';
 import secretService from './secret.service.js';
 import { LogActionEnum } from '../models/Telemetry.js';
 import encryptionManager from '../utils/encryption.manager.js';
 import errorManager, { ErrorSourceEnum } from '../utils/error.manager.js';
+import { getGlobalOAuthCallbackUrl } from '../utils/utils.js';
 
-import type { Orchestrator } from '../index.js';
+import type { Orchestrator } from '../clients/orchestrator.js';
 import type { Knex } from '@nangohq/database';
 import type { DBEnvironment, DBEnvironmentVariable, SdkLogger } from '@nangohq/types';
